attendance/requirements.txt
pymysql
pyyaml
elastic-apm[flask]
flask
cryptography
gunicorn
